+++ src/App.tsx
@@ -3,7 +3,7 @@ import "./App.css";
 import GameBoard from "./components/GameBoard";
 import { useState } from "react";
 import { GameConfig } from "@/state/types";
-import { GameStateProvider } from "@/context/GameStateProvider";
+import { GameStateProvider } from "@/context/game-state/provider";
 import { TimeMachine } from "@/components/TimeMachine";
 
 const App = () => {
