+++ src/components/CardAbilityTiles.tsx
@@ -1,7 +1,7 @@
 import GameElement from "@/components/GameElement";
 import { getHighlightTextureAssetPath } from "@/components/utils";
 import { abilityOffset } from "@/constants/gameBoard";
-import { useGameState } from "@/context/GameStateProvider";
+import { useGameState } from "@/context/game-state/provider";
 import { useSRGBTexture } from "@/hooks/useSRGBTexture";
 import { Coordinate } from "@/state/types";
 import { useTexture } from "@react-three/drei";
